requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
billiard==3.6.4.0
boto3==1.26.26
botocore==1.29.26
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.3
Django==4.1.3
django-environ==0.9.0
django-jazzmin==2.6.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
eventlet==0.33.2
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.50.0
grpcio-status==1.50.0
gTTS==2.3.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
msgpack==1.0.4
Pillow==9.3.0
prompt-toolkit==3.0.33
proto-plus==1.22.1
protobuf==4.21.9
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.6
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
